move build/watch out into separate modules
