+++ bin/src/run/index.js
@@ -2,9 +2,9 @@ import path from 'path';
 import { realpathSync } from 'fs';
 import * as rollup from 'rollup';
 import relative from 'require-relative';
-import chalk from 'chalk';
-import { handleWarning, handleError, stderr } from './logging.js';
-import SOURCEMAPPING_URL from './sourceMappingUrl.js';
+import { handleWarning, handleError } from '../logging.js';
+import build from './build.js';
+import watch from './watch.js';
 
 import { install as installSourcemapSupport } from 'source-map-support';
 installSourcemapSupport();
@@ -192,96 +192,8 @@ function execute ( options, command ) {
 	});
 
 	if ( command.watch ) {
-		if ( !options.entry || ( !options.dest && !options.targets ) ) {
-			handleError({
-				code: 'WATCHER_MISSING_INPUT_OR_OUTPUT',
-				message: 'must specify --input and --output when using rollup --watch'
-			});
-		}
-
-		try {
-			const watch = relative( 'rollup-watch', process.cwd() );
-			const watcher = watch( rollup, options );
-
-			watcher.on( 'event', event => {
-				switch ( event.code ) {
-					case 'STARTING': // TODO this isn't emitted by newer versions of rollup-watch
-						stderr( 'checking rollup-watch version...' );
-						break;
-
-					case 'BUILD_START':
-						stderr( 'bundling...' );
-						break;
-
-					case 'BUILD_END':
-						stderr( 'bundled in ' + event.duration + 'ms. Watching for changes...' );
-						break;
-
-					case 'ERROR':
-						handleError( event.error, true );
-						break;
-
-					default:
-						stderr( 'unknown event', event );
-				}
-			});
-		} catch ( err ) {
-			if ( err.code === 'MODULE_NOT_FOUND' ) {
-				handleError({
-					code: 'ROLLUP_WATCH_NOT_INSTALLED',
-					message: 'rollup --watch depends on the rollup-watch package, which could not be found. Install it with npm install -D rollup-watch'
-				});
-			}
-
-			handleError( err );
-		}
+		watch( options );
 	} else {
-		bundle( options ).catch( handleError );
+		build( options ).catch( handleError );
 	}
-}
-
-function clone ( object ) {
-	return assign( {}, object );
-}
-
-function assign ( target, source ) {
-	Object.keys( source ).forEach( key => {
-		target[ key ] = source[ key ];
-	});
-	return target;
-}
-
-function bundle ( options ) {
-	return rollup.rollup( options )
-		.then( bundle => {
-			if ( options.dest ) {
-				return bundle.write( options );
-			}
-
-			if ( options.targets ) {
-				let result = null;
-
-				options.targets.forEach( target => {
-					result = bundle.write( assign( clone( options ), target ) );
-				});
-
-				return result;
-			}
-
-			if ( options.sourceMap && options.sourceMap !== 'inline' ) {
-				handleError({
-					code: 'MISSING_OUTPUT_OPTION',
-					message: 'You must specify an --output (-o) option when creating a file with a sourcemap'
-				});
-			}
-
-			return bundle.generate(options).then( ({ code, map }) => {
-				if ( options.sourceMap === 'inline' ) {
-					code += `\n//# ${SOURCEMAPPING_URL}=${map.toUrl()}\n`;
-				}
-
-				process.stdout.write( code );
-			});
-		})
-		.catch( handleError );
 }
